抽奖参数配置化

+++ hk-medicine-api/medicine-music-api/src/main/java/com/hk/music/api/service/UserRaffleDetailsService.java
@@ -24,7 +24,7 @@ public interface UserRaffleDetailsService {
      *
      * @param ethAmountTotal
      */
-    void openUserRaffle(BigDecimal ethAmountTotal);
+    void openUserRaffle(double ethAmountTotal);
 
 
     /**
